refactor palette setting to use events
